fundamentals
Here I likened a stringed nodes to stringed coins. The flipping of coins is rife for our purpose. We need to simulate the flipping of four coins in each string. 
The will be done for 4 different strings. We are looking at 256 outcomes. How they will be used will be explained.